Commit de un montón de cosas (muchas seguro al pedo) tratando de arreglar el problema de la creación de rutas.

+++ AerolineaFrba/AerolineaFrba/Abm Rol/CargarRolForm.cs
@@ -61,7 +61,7 @@ namespace AerolineaFrba.Abm_Rol
                 {
                     // insert
                     this.rol = new Rol();
-                    this.rol.Id = Convert.ToInt16(this.rolesTableAdapter.rolInsertar(this.descripcion.Text));
+                    this.rol.Id = Convert.ToInt32(this.rolesTableAdapter.rolInsertar(this.descripcion.Text));
                 }
 
                 foreach (object itemChecked in this.checkedListBox1.CheckedItems)
